blog-editor-api/requirements.txt
# Blog Editor API Dependencies

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyGithub>=2.1.0
httpx>=0.26.0
python-multipart>=0.0.6
Pillow>=10.2.0
PyYAML>=6.0.1
